+++ components/quote-carousel/QuoteCarousel.style.js
@@ -6,8 +6,9 @@ import {
   FadeIn,
   FadeOut,
 } from '../animation-base-components/AnimationBaseComponents';
+import { motion } from 'framer-motion';
 
-const ContentWrapper = styled.div`
+const ContentWrapper = styled(motion.div)`
   width: 55em;
   margin: auto;
   display: flex;
