+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "axios": "^0.21.1",
     "babel-plugin-styled-components": "^1.12.0",
+    "framer-motion": "^4.1.13",
     "intersection-observer": "^0.12.0",
     "moment": "^2.29.1",
     "next": "10.0.7",
